Move `article.ts` into `types` directory
