+++ dotcom-rendering/src/components/ArticlePage.tsx
@@ -1,12 +1,12 @@
 import { Global } from '@emotion/react';
 import { ArticleDesign, type ArticleFormat } from '@guardian/libs';
 import { StrictMode } from 'react';
-import type { Article } from '../article';
 import { DecideLayout } from '../layouts/DecideLayout';
 import { buildAdTargeting } from '../lib/ad-targeting';
 import { rootStyles } from '../lib/rootStyles';
 import { filterABTestSwitches } from '../model/enhance-switches';
 import type { NavType } from '../model/extract-nav';
+import type { Article } from '../types/article';
 import type { RenderingTarget } from '../types/renderingTarget';
 import { AlreadyVisited } from './AlreadyVisited.importable';
 import { BrazeMessaging } from './BrazeMessaging.importable';
