+++ dotcom-rendering/src/layouts/FullPageInteractiveLayout.tsx
@@ -6,7 +6,6 @@ import {
 	palette as sourcePalette,
 	until,
 } from '@guardian/source/foundations';
-import type { Article } from '../article';
 import {
 	adContainerStyles,
 	MobileStickyContainer,
@@ -28,6 +27,7 @@ import { decideLanguage, decideLanguageDirection } from '../lib/lang';
 import { renderElement } from '../lib/renderElement';
 import type { NavType } from '../model/extract-nav';
 import { palette as themePalette } from '../palette';
+import type { Article } from '../types/article';
 import type { ServerSideTests, Switches } from '../types/config';
 import type { FEElement } from '../types/content';
 import { interactiveGlobalStyles } from './lib/interactiveLegacyStyling';
