+++ dotcom-rendering/src/layouts/InteractiveLayout.tsx
@@ -9,7 +9,6 @@ import {
 import { Hide } from '@guardian/source/react-components';
 import { StraightLines } from '@guardian/source-development-kitchen/react-components';
 import React from 'react';
-import type { Article } from '../article';
 import { AdPortals } from '../components/AdPortals.importable';
 import { AdSlot, MobileStickyContainer } from '../components/AdSlot.web';
 import { AppsFooter } from '../components/AppsFooter.importable';
@@ -48,6 +47,7 @@ import { getContributionsServiceUrl } from '../lib/contributions';
 import { decideTrail } from '../lib/decideTrail';
 import type { NavType } from '../model/extract-nav';
 import { palette as themePalette } from '../palette';
+import type { Article } from '../types/article';
 import type { RenderingTarget } from '../types/renderingTarget';
 import {
 	interactiveGlobalStyles,
