+++ dotcom-rendering/src/layouts/NewsletterSignupLayout.tsx
@@ -19,7 +19,6 @@ import {
 	SvgGuardianLogo,
 } from '@guardian/source/react-components';
 import { StraightLines } from '@guardian/source-development-kitchen/react-components';
-import type { Article } from '../article';
 import { AdSlot, MobileStickyContainer } from '../components/AdSlot.web';
 import { ArticleHeadline } from '../components/ArticleHeadline';
 import { Carousel } from '../components/Carousel.importable';
@@ -46,6 +45,7 @@ import { decideTrail } from '../lib/decideTrail';
 import { isValidUrl } from '../lib/isValidUrl';
 import type { NavType } from '../model/extract-nav';
 import { palette as themePalette } from '../palette';
+import type { Article } from '../types/article';
 import { BannerWrapper, Stuck } from './lib/stickiness';
 
 type Props = {
