+++ dotcom-rendering/src/server/handler.article.apps.ts
@@ -1,5 +1,5 @@
 import type { RequestHandler } from 'express';
-import { enhanceArticleType } from '../article';
+import { enhanceArticleType } from '../types/article';
 import { makePrefetchHeader } from './lib/header';
 import { recordTypeAndPlatform } from './lib/logging-store';
 import { renderArticle } from './render.article.apps';
